AsyncStoreStressTest: support multiple test runs
DummyInMenoryCacheStore keeps track of all stores by name in a static
variable. If the stress test is run twice, the named memory stores are
already filled with data and validation fails at the end. Clear the stores
so that the test can be executed several times (e.g. with different thread
counts).

Signed-off-by: Karsten Blees <blees@dcon.de>