AsyncStore2: allow the modificationQueueSize feature to be switched off
The modificationQueueSize feature doesn't work properly in the original
AsyncStore. Allow the feature to be switched off in the new AsyncStore
implementation (by setting modificationQueueSize == 0) to get meaningful
performance measurements.

With modificationQueueSize == 0, it is no longer possible to calculate the
batch size of AsyncStoreProcessor threads in advance. Rewrite the thread
setup code to distribute entries more evenly across available threads.

Signed-off-by: Karsten Blees <blees@dcon.de>