ISPN-5559 AsyncCacheWriter: don't use timeouts when waiting for AsyncStoreProcessors
This reverts commit 59819640 "ISPN-2476 Worker threads latch awaits should
be timed"

This fixes the "AsyncCacheWriter is dead!" bug reported on the forum [1]
and mailing list [2].

With assertions disabled, waiting with timeout causes deadlock, because the
coordinator thread exits with AssertionError and thus can no longer process
changes.

With assertions enabled, waiting with timeout causes unbounded use of
memory, as the coordinator thread will allocate an unbounded number of
AsyncStoreProcessor objects without waiting for data to be written to the
back-end store.

Luckily, the AsyncStoreProcessor thread pool has been limited to a single
thread (see ISPN-3532), so this bug could not cause any data corruption due
to out-of-order writes (the coordinator thread only guaratees disjoint keys
for two states, i.e. only two states can be processed in parallel).

[1] https://developer.jboss.org/thread/237159
[2] http://lists.jboss.org/pipermail/infinispan-dev/2014-November/015707.html

Signed-off-by: Karsten Blees <blees@dcon.de>