ISPN-3532 AsyncCacheWriter: fix multi-threaded back-end writes
ThreadPoolExecutor only creates more than corePoolSize threads if the
workQueue is full. With an unbounded workQueue and corePoolSize == 0, the
thread pool is always single threaded.

Fix this by setting corePoolSize to the configured number of threads.

Additionally, allow core pool threads to timeout, so that we don't use
threads if the cache is idle.

Writing State data to the back-end store can now be completed out-of-order
(i.e. the current state's processors may finish before the tail state's
processors). For State.get() to work correctly, AsyncStoreProcessor.run()
must not disconnect State objects with an active next state.

Signed-off-by: Karsten Blees <blees@dcon.de>