ISPN-5563 AsyncStoreTest: test more thoroughly that load() is in sync with changes
That load() sometimes doesn't see changes that have been written to the
AsyncCacheWriter (former AsyncStore) was the root cause of ISPN-1174.

Fix the tests that still assume that "waiting" on changes to appear in
the CacheLoader is correct behaviour.

Testing the basic operations with 1,000,000 instead of just 1,000 entries
also shows that the issue is still not completely fixed with redesign of
the AsyncStore (ISPN-2293).

Signed-off-by: Karsten Blees <blees@dcon.de>